JS-FIX_GlobalVariable=>Class_and_others
add Class obj
js modules split
ignore
