+++ src/js/index.js
@@ -1,4 +1,8 @@
 import { stAnime } from "./startAnime.js";
+import { Player } from "./class.js";
+import { dakuten } from "./dakuten.js";
+import { HBG_menu, onlyPortrait } from "./others.js";
+
 
 // ひらがな５０音元配列( Length => 46 )
 const data = [
@@ -41,24 +45,10 @@ let firstPlayer;
 let secondPlayer;
 // (プレーヤー名を代入)
 
-
-// プレーヤーの定義
-let player1 = {
-  name: 'とり',
-  avatar: './image/tori.png',
-  handCards: [],
-  changesLeft: 4,
-  description: '',
-  winner: false
-}
-let player2 = {
-  name: 'いぬ',
-  avatar: './image/inu.png',
-  handCards: [],
-  changesLeft: 4,
-  description: '',
-  winner: false
-}
+// // プレーヤー
+// プレーヤーインスタンス作成
+let player1 = new Player( 'とり', 'image/tori.png', [], 4, '',  false );
+let player2 = new Player( 'いぬ', 'image/inu.png', [], 4, '',  false );
 
 // 行動中のプレイヤー
 let currentPlayer = {
@@ -67,8 +57,7 @@ let currentPlayer = {
   avatar: '', // url
   handCards: [],
   description: '',
-  changesLeft: 4,
-  finishedPlayer: 0
+  changesLeft: 4
 }
 
 let GameState = {
@@ -81,6 +70,7 @@ let GameState = {
 
 document.addEventListener('DOMContentLoaded', () => {
   // geme
+  onlyPortrait();
   pushStartBtn();
   cardChange();
   moreCardChangeDone();
@@ -107,7 +97,7 @@ document.addEventListener('DOMContentLoaded', () => {
 window.addEventListener('scroll', () => {
   let timeoutId;
   if ( timeoutId ) return;
-  timeoutId = setTimeout( function () {
+  timeoutId = setTimeout( () => {
     timeoutId = 0;
   GoToNextStage()}, 2000);
 });
@@ -294,7 +284,6 @@ const setFirstPlayerTori = () => {
   currentPlayer.avatar = './image/tori.png';
   currentPlayer.handCards = firstPlayerArray;
   currentPlayer.description = '';
-  currentPlayer.acted = false;
 
   firstPlayer = player1.name;
   secondPlayer = player2.name;
@@ -316,7 +305,6 @@ const setFirstPlayerInu = () => {
   currentPlayer.avatar = './image/inu.png';
   currentPlayer.handCards = firstPlayerArray;
   currentPlayer.description = '';
-  currentPlayer.acted = false;
 
   firstPlayer = player2.name;
   secondPlayer = player1.name;
@@ -657,7 +645,7 @@ const sortWindowAppearance = () => {
   const toSortWindowBtn = document.getElementById('toSortWindowBtn');
 
   toSortWindowBtn.onclick = () => {
-    // gameStage更新
+    // GameStage更新
     GameState.stage = 'sortStage'
     GameState.reScrollPoint = 'sort'
 
@@ -765,7 +753,7 @@ const toDescription = () => {
     //ボタンの非活性化
     isDisabled(toDescriptionBtn,true);
 
-    // gameStage更新
+    // GameStage更新
     GameState.stage = 'descriptionStage'
     GameState.reScrollPoint = 'description'
 
@@ -806,7 +794,7 @@ const backToSort = () => {
     const toDescriptionBtn = document.getElementById('toDescriptionBtn')
     removeDisabled(toDescriptionBtn);
 
-    // gameStage更新
+    // GameStage更新
     GameState.stage = 'sortStage'
     GameState.reScrollPoint = 'sort'
 
@@ -1033,12 +1021,10 @@ const finalConfirmModal = () => {
 
       // console.log('Player2のターンへ');
 
-      // // 画面を一番上元に戻す
-      // scrollTop(500);
 
       // // 重さを考慮して↓に差し替え
 
-      // gameStage更新
+      // GameStage更新
       GameState.stage = 'firstStage'
       GameState.reScrollPoint = 'first'
 
@@ -1137,7 +1123,7 @@ const finalConfirmModal = () => {
       // console.log(player1);
       // console.log(player2);
 
-      // gameStage更新
+      // GameStage更新
       GameState.stage = 'resultStage'
       GameState.reScrollPoint = 'resultStage'
 
@@ -1156,20 +1142,17 @@ const finalConfirmModal = () => {
 const dataMove = () => {
   const description_Ele = document.getElementById('descriptionDisplayAreaB');
   currentPlayer.description = description_Ele;
-  currentPlayer.acted = true;
 
   if ( currentPlayer.player == player1 ) {
 
     player1.changesLeft = currentPlayer.changesLeft, // 0
     player1.handCards = currentPlayer.handCards,
-    player1.description = currentPlayer.description,
-    player1.acted =  currentPlayer.acted
+    player1.description = currentPlayer.description
   }
   if ( currentPlayer.player == player2 ) {
     player2.changesLeft = currentPlayer.changesLeft, // 0
     player2.handCards = currentPlayer.handCards,
-    player2.description = currentPlayer.description,
-    player2.acted =  currentPlayer.acted
+    player2.description = currentPlayer.description
   }
 
 };
@@ -1227,28 +1210,10 @@ const setDisplayDefault = () => {
   descriptionWriteArea.value = ''
 }
 
-// //画面をスクロールで一番上に戻す(スピード変更可)
-//   function scrollTop(duration) {
-//     let currentY = window.pageYOffset;
-//     let step = duration/currentY > 1 ? 10 : 100;
-//     let timeStep = duration/currentY * step;
-//     let intervalID = setInterval(scrollUp, timeStep);
-
-//     function scrollUp(){
-//       currentY = window.pageYOffset;
-//       if(currentY === 0) {
-//           clearInterval(intervalID);
-//       } else {
-//           scrollBy( 0, -step );
-//       }
-//     }
-//   }
-
 
 const setSecondPlayer = () => {
 // 次プレイヤー用の値をオブジェクトにセットする
   if (currentPlayer.player == player1) {
-    player1.acted = true;
 
     currentPlayer.name = 'いぬ';
     currentPlayer.player = player2;
@@ -1256,14 +1221,12 @@ const setSecondPlayer = () => {
     currentPlayer.handCards = secondPlayerArray;
     currentPlayer.description = '';
     currentPlayer.changesLeft = 4;
-    currentPlayer.acted = false;
     GameState.finishedPlayer = 1;
     GameState.stage = 'firstStage';
     GameState.reScrollPoint = 'first';
 
   }
   else if (currentPlayer.player == player2) {
-    player2.acted = true;
 
     currentPlayer.name = 'とり';
     currentPlayer.player = player1;
@@ -1271,7 +1234,6 @@ const setSecondPlayer = () => {
     currentPlayer.handCards = secondPlayerArray;
     currentPlayer.description = '';
     currentPlayer.changesLeft = 4;
-    currentPlayer.acted = false;
     GameState.finishedPlayer = 1;
     GameState.stage = 'firstStage';
     GameState.reScrollPoint = 'first';
@@ -1335,6 +1297,7 @@ const openResultModal = () => {
         isHidden(resultModalSecond); // 開
         isHidden(lastResult); //開
         lastResult.innerHTML = `${firstPlayer}さんの勝ち！`
+        console.log(player1,player2)
       }
       // 1:1
       secondPlayerResultB.onclick = () => {
@@ -1370,219 +1333,9 @@ const openResultModal = () => {
 }
 
 
-// *****濁点*****
-
-const dakuten = () => {
-  const sortResetBtn = document.getElementById('sortResetBtn');
-  const dakutenPopTarget = document.getElementById('dakutenPopTarget');
-  const CP_HandCards = document.getElementById('currentPlayerSortAfter');
-  const CP_HandCards_Elem_Li = document.getElementById('currentPlayerSortAfter').getElementsByTagName('li');
-  const CP_HandCardsArray = Array.prototype.slice.call(CP_HandCards_Elem_Li);
-
-  // カードをクリックしたら、文字を判定する + 要素の情報を次の関数に渡す
-  CP_HandCardsArray.forEach (function FxA(ele,index) {
-    ele.addEventListener('click', () => {
-
-      // 出ているポップの(要素)を消す
-      while (dakutenPopTarget.firstChild) dakutenPopTarget.removeChild(dakutenPopTarget.firstChild);
-
-      // カードの文字の判定して、元文字と候補を次の関数に送る
-      switch ( ele.innerHTML ) {
-        case 'あ': openPop_dakuten('あ','','','ぁ',index);
-          break;
-        case 'い': openPop_dakuten('い','','','ぃ',index);
-          break;
-        case 'う': openPop_dakuten('う','ゔ','','ぅ',index);
-          break;
-        case 'え': openPop_dakuten('え','','','ぇ',index);
-          break;
-        case 'お': openPop_dakuten('お','','','ぉ',index);
-          break;
-        case 'か': openPop_dakuten('か','が','','',index);
-          break;
-        case 'き': openPop_dakuten('き','ぎ','','',index);
-          break;
-        case 'く': openPop_dakuten('く','ぐ','','',index);
-          break;
-        case 'け': openPop_dakuten('け','げ','','',index);
-          break;
-        case 'こ': openPop_dakuten('こ','ご','','',index);
-          break;
-        case 'さ': openPop_dakuten('さ','ざ','','',index);
-          break;
-        case 'し': openPop_dakuten('し','じ','','',index);
-          break;
-        case 'す': openPop_dakuten('す','ず','','',index);
-          break;
-        case 'せ': openPop_dakuten('せ','ぜ','','',index);
-          break;
-        case 'そ': openPop_dakuten('そ','ぞ','','',index);
-          break;
-        case 'た': openPop_dakuten('た','だ','','',index);
-          break;
-        case 'ち': openPop_dakuten('ち','ぢ','','',index);
-          break;
-        case 'つ': openPop_dakuten('つ','づ','','っ',index);
-          break;
-        case 'て': openPop_dakuten('て','で','','',index);
-          break;
-        case 'と': openPop_dakuten('と','ど','','',index);
-          break;
-        case 'は': openPop_dakuten('は','ば','ぱ','',index);
-          break;
-        case 'ひ': openPop_dakuten('ひ','び','ぴ','',index);
-          break;
-        case 'ふ': openPop_dakuten('ふ','ぶ','ぷ','',index);
-          break;
-        case 'へ': openPop_dakuten('へ','べ','ぺ','',index);
-          break;
-        case 'ほ': openPop_dakuten('ほ','ぼ','ぽ','',index);
-          break;
-        case 'や': openPop_dakuten('や','','','ゃ',index);
-          break;
-        case 'ゆ': openPop_dakuten('ゆ','','','ゅ',index);
-          break;
-        case 'よ': openPop_dakuten('よ','','','ょ',index);
-          break;
-        case 'ー': openPop_dakuten('ー','〜','','-',index);
-          break;
-      }
-    })
-  })
-
-  function openPop_dakuten (moto,ten,maru,komoji,index) {
-    // *要素の生成*
-    // div 入れ物
-    let Div = document.createElement('div');
-    Div.className = 'p-pop__dakuten';
-    // 元の文字
-    let Moto = document.createElement('p')
-    Moto.innerHTML = moto
-    Div.appendChild(Moto)
-    Moto.setAttribute('id', 'optionDefault');
-    // ⇆
-    let Arrow = document.createElement('p')
-    // Arrow.innerHTML = '<i class="fas fa-long-arrow-alt-right"></i>'
-    Arrow.innerHTML = '<i class="fas fa-exchange-alt"></i>'
-    Div.appendChild(Arrow)
-    // 点
-    if ( ten != '' ){
-      let Ten = document.createElement('p')
-      Ten.innerHTML = ten
-      Div.appendChild(Ten)
-      Ten.setAttribute('id', 'optionA');
-    }
-    // 丸
-    if ( maru != '' ) {
-      let Maru = document.createElement('p')
-      Maru.innerHTML = maru
-      Div.appendChild(Maru)
-      Maru.setAttribute('id', 'optionB');
-    }
-    // 小文字
-    if ( komoji != '' ) {
-      let Komoji = document.createElement('p')
-      Komoji.innerHTML = komoji
-      Div.appendChild(Komoji)
-      Komoji.setAttribute('id', 'optionC');
-    }
-
-    // 完成した要素を挿入 '.p-pop__dakuten ＞ p*4'
-    dakutenPopTarget.appendChild(Div);
-
-    // 各選択肢をクリックした時
-
-    choiceOptions(optionDefault,index);
-    if ( ten != '' ) {
-      choiceOptions(optionA,index);
-    }
-    if ( maru != '' ) {
-      choiceOptions(optionB,index);
-    }
-    if ( komoji != '' ) {
-      choiceOptions(optionC,index);
-    }
-
-  }
-
-  //  ポップの文字をクリックした時にカードの文字を変える
-  function choiceOptions (target,index) {
-    target.addEventListener('click',() => {
-      CP_HandCards.children[index].innerHTML = target.innerHTML;
-    })
-  }
-
-  // 「もう一回並べ換える」ボタンを押した時、出ているポップを消す。
-  sortResetBtn.addEventListener('click', () => {
-    while (dakutenPopTarget.firstChild) dakutenPopTarget.removeChild(dakutenPopTarget.firstChild);
-  })
-
-}
-
-// *****濁点*****
-
-
 
 // 勝ち数の記録を表示？
 // ネット対戦…？
 // Vueで作り直す…？
 
-
-
-
-
-
-// ********** HBG **********
-const HBG_menu = () =>{
-
-  const hamburgerBtn = document.getElementById('hamburgerBtn');
-  const navList = document.getElementById('navList');
-
-  hamburgerBtn.addEventListener('click', function() {
-    setAriaExpanded(hamburgerBtn);
-    setAriaExpanded(navList);
-  })
-
-  const toRules = document.getElementById('toRules');
-  const Rules = document.getElementById('Rules');
-  const closeRules = document.getElementById('closeRules');
-  const toRequirements = document.getElementById('toRequirements');
-  const Requirements = document.getElementById('Requirements');
-  const closeRequirements = document.getElementById('closeRequirements');
-
-  toRules.onclick = () => {
-    setAriaExpanded(Rules);
-    setAttr(Requirements, false)
-    // setAttr(hamburgerBtn, false);
-    // setAttr(navList, false);
-  };
-  closeRules.onclick = () => {
-    setAriaExpanded(Rules);
-  };
-  toRequirements.onclick = () => {
-    setAriaExpanded(Requirements);
-    setAttr(Rules, false)
-    // setAttr(hamburgerBtn, false);
-    // setAttr(navList, false);
-  };
-  closeRequirements.onclick = () => {
-    setAriaExpanded(Requirements);
-  };
-
-
-}
-
-// スマホの横置き時、「縦でお願いします」のダイアログを出す。
-window.onorientationchange = () => {
-  switch ( window.orientation ) {
-    case 0:
-      break;
-    case 90:
-      alert('スマホでは画面を縦にしてプレイしてください');
-      break;
-    case -90:
-      alert('スマホでは画面を縦にしてプレイしてください２');
-      break;
-  }
-}
